resumeanalyzer/requirements.txt
pydantic==1.8.2
langchain-g4f==0.0.1
g4f==0.0.2
python-dotenv==0.21.0
agency-swarm==0.1.0
